Change to generate_sorted_append_paths
It's a renaming of generate_mergeappend_paths which will also created
sorted Append paths.

Compute the pathkeys for the related rel at that time, and change the
generate_pathkeys_for_partitioned_table to better suit this new
approache.  The relation is not needed anymore, use only fields in
RelOptInfo.

Still needs to check if the partition pathkeys match and actually create
the needed paths.