Cleanup cached plan query strings
When a multi statements query string contains a query that is cached (for
instance if one of the statements is a PREPARE statement), the whole query
string was saved.  This is wasteful as that string can be artitrarily big, but
it can also confusing as some other parts like pg_prepared_statements will
output the saved query string as-is.

This commit changes this behavior and only stores the part of the query string
that correspond to any given cached plan, similarly to how it's already done in
pg_stat_statements.